refactor: extract hierarchy ops into wvImportHierarchyController
Move undoLastOperation out of the worldViewImportController monolith
into a dedicated wvImportHierarchyController file. Reads undoEntries
from wvImportSharedState (introduced in PR-14).

Pure refactor — no behavior change, byte-identical function body.
Continues the spine domain-split (ADR-0009).
